splash screen smooth transistion
