+++ app/home/page.tsx
@@ -1,0 +1,13 @@
+import Navbar from '@/components/Navbar';
+import Hero from '@/components/Hero';
+import WelcomeSection from '@/components/WelcomeSection';
+
+export default function HomePage() {
+  return (
+    <>
+      <Navbar />
+      <Hero />
+      <WelcomeSection />
+    </>
+  );
+}
